tests: working out with individual method tests
