+++ Tests/AuthenticationTests/Mocks/Errors/MissingTokenRequest/MockAuthorizationResponse_MissingTokenRequest.swift
@@ -1,0 +1,10 @@
+import AppAuthCore
+
+class MockAuthorizationResponse_MissingTokenRequest: OIDAuthorizationResponse {
+    override func tokenExchangeRequest(
+        withAdditionalParameters additionalParameters: [String: String]?,
+        additionalHeaders: [String: String]?
+    ) -> OIDTokenRequest? {
+        nil
+    }
+}
